student-performance/requirements.txt
flask
mysql-connector-python
pandas
matplotlib
werkzeug
openpyxl
